refactor: simplify base input and fix border
