+++ src/shared/components/ui/input/password-input.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { BaseInput } from 'shared/styling/components/text-input.emotion';
+import { BaseInput } from 'shared/styling/components/input.emotion';
 
 export const PasswordInput = (
   props: React.InputHTMLAttributes<HTMLInputElement>
